+++ main/protocols/pylon/pylon_rs485_bridge.h
@@ -18,6 +18,7 @@ void pylonRs485BridgeStop(void);
 #ifdef HOST_TEST
 bool pylonRs485BridgeBuildSyntheticInfo61ForTest(char *out, size_t outSize);
 bool pylonRs485BridgeBuildSyntheticInfo63ForTest(char *out, size_t outSize);
+void pylonRs485BridgeRefreshSyntheticCacheForTest(void);
 bool pylonRs485BridgeProbeShouldWaitForQuietForTest(uint8_t mode,
                                                     int64_t nowUs,
                                                     int64_t lastBmsTrafficUs,